confirm dialogue works but ugly ;

+++ client/src/components/DeleteExpense.js
@@ -4,6 +4,7 @@ import { ALL_EXPENSES } from '../graphQL/queries'
 import { useMutation } from '@apollo/react-hooks'
 import { Button } from 'semantic-ui-react'
 import { withRouter } from 'react-router'
+import { Confirmation } from '../components'
 
 const DeleteExpense = props => {
   const [deleteExpense] = useMutation(DELETE_EXPENSE, {
@@ -11,13 +12,15 @@ const DeleteExpense = props => {
   })
   const handleDelete = async id => {
     console.log('handle delete firing', id)
-    try {
-      await deleteExpense({
-        variables: { id }
-      })
-      props.history.push('/expenses')
-    } catch (error) {
-      console.log('something went wrong with deleting expense', error)
+    if (window.confirm('are you sure???')) {
+      try {
+        await deleteExpense({
+          variables: { id }
+        })
+        props.history.push('/expenses')
+      } catch (error) {
+        console.log('something went wrong with deleting expense', error)
+      }
     }
   }
   return (
